# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Self-Service-Totem MAKING API",
        "description": "Documentation for the kitchen to prepare and deliver orders",
        "version": "1.0.0",
        "contact": {
            "name": "API Support Team",
            "url": "https://github.com/fiap-soat-sst/sst-ms-making/issues"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base path"
        }
    ],
    "paths": {
        "/kitchen/order/:id": {
            "get": {
                "summary": "get one order that is being prepared",
                "operationId": "listOrders",
                "description": "Get orders",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Orders found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderOutputDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: Unable to find orders"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/kitchen/order": {
            "post": {
                "summary": "Create a new order to be prepared",
                "tags": ["Orders"],
                "operationId": "createOrder",
                "description": "Create a new order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderInputDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Order created successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: Missing or invalid parameters"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/kitchen/list-all": {
            "get": {
                "summary": "List all orders",
                "description": "Get all orders",
                "operationId": "listAllOrders",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Orders found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderOutputDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: Unable to find orders"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/kitchen/order/update-status/preparing": {
            "put": {
                "summary": "Prepare an order",
                "description": "Prepare an order",
                "operationId": "prepareOrder",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Order updated successfully to Preparing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated successfully"
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/kitchen/order/update-status/finishing": {
            "put": {
                "summary": "Finish preparing an order",
                "description": "Finish preparing an order",
                "operationId": "finishPreparingOrder",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Order updated successfully to Ready",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated successfully"
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/kitchen/order/update-status/delivered": {
            "put": {
                "summary": "Finish an order",
                "description": "Finish an order",
                "operationId": "finishOrder",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Order updated successfully to Finished",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated successfully"
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OrderInputDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "OrderUpdateDTO": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "Recebido",
                            "Em preparação",
                            "Pronto",
                            "Finalizado"
                        ]
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "OrderOutputDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "customer": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "productId": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "closed": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders"
        }
    ]
}